+++ src/batch/src/executor2/mod.rs
@@ -19,7 +19,7 @@ mod filter;
 mod hash_agg;
 mod insert;
 mod limit;
-mod projection;
+mod project;
 mod trace;
 mod values;
 pub use delete::*;
@@ -28,7 +28,7 @@ use futures::stream::BoxStream;
 pub use hash_agg::*;
 pub use insert::*;
 pub use limit::*;
-pub use projection::*;
+pub use project::*;
 use risingwave_common::array::DataChunk;
 use risingwave_common::catalog::Schema;
 use risingwave_common::error::Result;
